type fix retriver to retriever. splitting files
