+++ backtest/template.py
@@ -4,7 +4,7 @@ import backtrader as bt
 import backtrader.feeds as btfeeds
 
 import pandas as pd
-import rba_tools.retriver.get_crypto_data as gcd
+import rba_tools.retriever.get_crypto_data as gcd
 import rba_tools.backtest.backtrader_extensions.strategies as rbsstrat
