+++ retriever/__init__.py
@@ -1,0 +1,1 @@
+from rba_tools.retriever import get_crypto_data as gcd
